Initial commit
A proof of concept wayland client with no dependencies.

* Open a unix socket to $XDG_RUNTIME_DIR/$WAYLAND_DISPLAY
* Bind to the wl_display interface
* Get registry
* Bind to a few relevant interfaces
* Create a pixel buffer in shared memory
* Share pixel buffer with compositor

In order to talk to the compositor, we need to serialize messages
according to the wayland protocol. This involves a 8 byte header along
with message specific data. The message data is interpreted according to
the types specified in the XML. See wl_reader and wl_writer for how the
types in protocol messages are interpreted on the wire

Wayland messages are defined in xml files, we parse these files and
generate zig structs out of them. The wl_reader/wl_writer iterate the
fields of the structs to figure out how to behave. This is a reasonably
simple codegen all things considered. This is not crazy, wayland-scanner
in the official wayland lib does the same thing but for C

XmlParser is stolen from sphmaps, but with some updates to read actual
xml content, not just names/attributes

Current state is pretty bare bones. We do not have a render loop, we
paint the pixel buffer one time, and ignore all non-essential requests
from the compositor.

The usage of the decoration manager may be a little surprising. Kwin
will not draw window decorations if we do not bind to that interface, so
we do

Implementation of file descriptor transfer is a little odd. At the time
of writing, zig does not have the cmsg API in the standard lib. We
experimented with writing it directly in zig, however using the C macros
is much more convenient, so we have some C glue code for use with
sendmsg

Some useful resources
* https://gaultier.github.io/blog/wayland_from_scratch.html
* https://wayland.freedesktop.org/docs/html/ch04.html
* https://wayland.app/protocols/