Log allowed dma buffer params
This was useful when debugging previous commit. Slight refactor required
around how we retrieve events to support mapping the format table from a
file descriptor. Previously we would just buffer events as a stream of
bytes, now we have to sync those bytes with ancillary data that comes in
with the headers.