Initial commit for a window manager
We're working on a window manager!

This is a first commit which can import a buffer from a client and
display it using KMS/DRM. Technically we probably can render 2 windows,
but they would display on top of eachother so it's effectively one. This
is very much a work in progress, but we have to start somewhere

The only expected case where this will work is with the integrated
wayland client. More work needs to be done to support real wayland apps

App is set up as an event loop that wakes up when
* A new wayland client connects
* A wayland client sends a message
* DRM notifies us of a page flip

Each subsystem goes through the central hub of CompositorState. The
eventual idea is that as data comes in, clients/input modify the
CompositorState. Each render frame we look at the current state and
render it as it makes sense

Rendering is done through one of 2 backends right now. The KMS/DRM
system which opens a /dev/dri/card and exports our GL buffers to it, and
the null backend is available for fallback/testing in an environment
where DRM is unavailable.

The flow of a client connection
* Client opens $XDG_RUNTIME_DIR/wayland-xxx
* sphwim registers the new connection with the event loop
* Wayland chatter happens. State lives in the Connection
* When a surface is ready for render, the buffer associated with it gets
  registered with CompositorState.
* On next vsync, we notice this buffer available to us, convert it into a
  texture, and render it
* At the end of each render we notify all clients that we want a new
  buffer
* A new buffer comes in at some point. Maybe within a vsync, maybe many
  vsyncs later
* We swap that buffer into compositor state and request a new one

Changes to existing codebase:
* When copying over the cmsg handling stuff, I wanted to do it without a
  C api boundary. Generate some constants which allow us to
  generate/receive cmsgs directly. Replace cmsg handling in client with
  new mechanism
* As a result of above, wlclient is a single file now and has been moved
  out of a single item folder
* wlgen has been updated to support both client and server bindings
* Correctly handle buffer params lifetimes in the client, as these were
  causing problems on the server
* Free wl_buffers from client