Proof of concept input handling
Hook up to libinput and udev to get input events from the system. Hook
this up to the compositor state and render a black box where the cursor
is on the screen

Note that at this point in time we are not doing proper VT management.
This means input events are still forwarded to the underlying VT AND
being processed by our window manager. This will later be the job of
systemd-logind, so we just ignore the problem for now