Automatic GPU selection
Remove hardcoded /dev/dri/card0 on both client and compositor. Wayland
compositors let clients know through the dmabuf feedback interface which
device should be used for rendering. Note that there is a per-surface
feedback mechanism as well, but EGL in mesa does not transfer buffers
between. We follow in their steps and just use whichever device is given
to use in the default feeback mechanism

On the server we've taken inspiration from kwin. Heuristically select a
GPU by counting how many monitors are plugged in, giving priority to
internal displays (e.g. on a laptop)

For now we just do one check on launch, but in the future we may have to
handle switching as monitors are added/removed