Fix virgl initialization
Virgl does not support creating resources with modifiers. We were
passing in modifiers because we were inspired by kmscube initialization
where they allow users to pass in a modifier on the command line. AFAICT
we do not care which modifiers are picked. Let the driver do what the
driver does and use the non-modifier API