Fix DRM initialization inconsistency
Previously we were calling getConnectorCurrent which passes parameters
to the drm GETCONNECTOR ioctl in such a way where it does not force
re-probe. In this mode, we also will only retrieve the first mode, which
may not be the preferred mode.

With some coarse debugging, it seems easy to end up in a case where
there are no preferred modes returned in this case. In qemu we can
reproduce the failure by flipping between view modes in the view
dropdown a couple times. I suspect that this changes the current mode
away from preferred, but that is not confirmed. I also suspect that on
initial run we may be getting lucky with a forced probe or something

Switch to using GetConnector, this has extra cost as we are not re-using
cached data, however since this is only done once on init it probably
doesn't matter