Reuse sphwim low level reader in client
Both the client and the server need to be able to read file descriptors
from the other. We don't want to write this twice. We experimented with
merging some of the higher level interfaces, but found in future commits
that having the ability for the client/server to diverge is nice.

Re-use the low level reader to get file descriptor passing, leave
everything else as is

Implementation details
* FdPool has been extracted from the wlio reader. What we're holding
  onto from the client's perspective (for now) is much more short lived.
  We only need the pooling behavior in the server. Also putting FdPool
  in wlio felt wrong, and pushing it into sphtud seemed unnecessary
* Related to the last bullet point, client events now need to be
  deinited to close their temporary file descriptors