Fix buffer corruption
While implementing wayland on wayland, we observe a new GPU
configuration. This is the first time we have seen GPU buffers created
on my actual desktop machine

On my desktop, both my radeon 6700xt and my integrated radeon GPU list
non-linear modifiers when querying buffer formats with EGL using
EGL_EXT_image_dma_buf_import_modifiers. This results in our GBM surfaces
being created with this modifier

In order to import non-linear textures, we need to import into a
GL_TEXTURE_EXTERNAL_OES texture, we are not currently doing this as we
are not a gles application. AFAICT when looking into this in the past,
it is not trivial to use this texture in the context of normal opengl.
Other window managers that we looked at use gles. Maybe one day we will
too (as android support will require us to switch), but for now we are
staying on desktop opengl

This limits us mostly to using EGL_EXT_image_dma_buf_import, which does
not define modifier import.

Importing with modifiers does not error, however usage of the imported
textures will result in what looks like corrupted data. The buffer
format is not respected by the glsl sampler

Limit modifier support to linear modifiers