Window ordering
Implement ordering of windows. Renderables are now sorted in storage
such that the most foreground window is the last in the list. This
allows new windows to be added cheaply, and foreground window removal to
also be relatively cheap. The idea thought here is that adds/removes are
more likely to be towards the front then the back of the stack

We also fix titlebar hit detection. We used to allow dragging a titlebar
through another window, now this is no longer the case