Switch renderable storage to multiarray
The original idea for renderables was that we could have an unsorted
list of them where ordering would be resolved on GPU during rendering.
As we implement window ordering, it becomes clear that we want the
windows sorted both for rendering and input hit testing

Rendering and input handling are happening many orders of magnitudes
more often than window reordering, so it's worth it (when we implement
ordering) to shuffle the renderables on reorder vs sorting them in a
first pass every input and rendering frame

While we're here, it also makes sense to keep different responsibilities
separated in memory for cache friendliness. Source info is only used for
frame updates, input testing will look at the position, but not the
render info, etc. etc.

This all points to a MultiArrayList

Usually we're a little hesitant to use the standard library types, as
they aren't super friendly to our allocators which are either arenas, or
only support large allocations. However in this case it makes sense
since we can just size our initial allocation to be ~= to 1 tiny page

We have 1 renderable list for the entire application, so over-allocating
by one tiny page is completely fine. Even keeping memory at peak usage
is also no big deal, as even with 100 windows the memory impact is
negligible (~60 bytes per renderable, so a page and a half)