Rename renderable -> window
Previously we were planning on keeping surfaces that were not
necessarily windows in the renderable list. With the addition of the
close button, we tie each surface back to the toplevel surface ID it
came from. We can no longer claim that this is anything but a window
abstraction

We can figure out how to handle non-window surfaces when we add them,
but for now we call them windows